chore(xlsx): extract chart legend host into chart/legend
Move every parseLegend / buildLegend / resolveLegend / normalizeLegend
helper from chart-reader.ts and chart-writer.ts (plus the LegendPos
type and ResolvedLegendEntry interface) into chart/legend.ts. The
unique clone-side normalize helpers (normalizeLegendBold /
normalizeLegendItalic / normalizeLegendUnderline /
normalizeLegendStrikethrough / normalizeLegendLayout /
normalizeLegendBorderWidth) move with them so chart-clone.ts can
import the shared per-field normalizers.

Pure relocation — JSDoc preserved verbatim, no behavior change. All
7362 tests still pass; typecheck clean.
