+++ src/xlsx/chart-reader.ts
@@ -82,6 +82,23 @@ import {
   parseTitleStrike,
   parseTitleUnderline,
 } from "./chart/title";
+import {
+  parseLegend,
+  parseLegendBold,
+  parseLegendBorderColor,
+  parseLegendBorderDash,
+  parseLegendBorderWidth,
+  parseLegendEntries,
+  parseLegendFillColor,
+  parseLegendFontColor,
+  parseLegendFontFamily,
+  parseLegendFontSize,
+  parseLegendItalic,
+  parseLegendLayout,
+  parseLegendOverlay,
+  parseLegendStrikethrough,
+  parseLegendUnderline,
+} from "./chart/legend";
 
 /** All chart-type element local names recognized by Excel. */
 const CHART_KIND_TAGS: ReadonlyMap<string, ChartKind> = new Map([
@@ -3719,583 +3736,6 @@ function parseSeriesStroke(ser: XmlElement): ChartLineStroke | undefined {
 }
 
 // ── Legend ────────────────────────────────────────────────────────
-
-/**
- * Map `<c:legend><c:legendPos val=".."/></c:legend>` to the writer-side
- * {@link ChartLegendPosition}. Returns `false` when `<c:delete val="1"/>`
- * is present (Excel's "no legend" state); returns `undefined` when the
- * chart has no `<c:legend>` element at all.
- */
-function parseLegend(chartEl: XmlElement): false | ChartLegendPosition | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-
-  // <c:delete val="1"/> means the chart explicitly suppresses the
-  // legend. Some Excel versions emit just an empty `<c:legend/>`
-  // followed by `<c:overlay/>` even when the legend is hidden, but
-  // `<c:delete val="1">` is the canonical "no legend" marker.
-  const del = findChild(legend, "delete");
-  if (del && readBoolVal(del.attrs.val) === true) return false;
-
-  const pos = findChild(legend, "legendPos");
-  if (!pos) {
-    // A legend element without legendPos is valid OOXML (Excel falls
-    // back to "right"). Surface "right" so the cloned chart preserves
-    // the visible-legend state.
-    return "right";
-  }
-  const val = pos.attrs.val;
-  if (typeof val !== "string") return "right";
-  switch (val) {
-    case "t":
-      return "top";
-    case "b":
-      return "bottom";
-    case "l":
-      return "left";
-    case "r":
-      return "right";
-    case "tr":
-      return "topRight";
-    default:
-      // Unknown legendPos values are dropped rather than fabricated.
-      return undefined;
-  }
-}
-
-/**
- * Pull `<c:legend><c:overlay val=".."/></c:legend>` off the chart. The
- * OOXML default `false` (the legend reserves its own slot, no overlap
- * with the plot area) collapses to `undefined` so absence and
- * `<c:overlay val="0"/>` round-trip identically through
- * {@link cloneChart} — only an explicit `<c:overlay val="1"/>` surfaces
- * `true`.
- *
- * The caller is expected to confirm a visible legend exists before
- * invoking this — `<c:overlay>` only renders when the legend is part of
- * the chart, so reading it from a chart that hides or omits the legend
- * would surface a flag with no on-screen effect.
- *
- * Accepts the OOXML truthy / falsy spellings (`"1"` / `"true"` / `"0"`
- * / `"false"`); unknown values and missing `val` attributes drop to
- * `undefined` rather than fabricate a flag Excel would not emit.
- */
-function parseLegendOverlay(chartEl: XmlElement): boolean | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  const overlay = findChild(legend, "overlay");
-  if (!overlay) return undefined;
-  const raw = overlay.attrs.val;
-  if (typeof raw !== "string") return undefined;
-  switch (raw) {
-    case "1":
-    case "true":
-      return true;
-    case "0":
-    case "false":
-      // OOXML default — collapse to undefined for symmetry with the
-      // writer's `legendOverlay` field.
-      return undefined;
-    default:
-      return undefined;
-  }
-}
-
-/**
- * Pull `<c:legend><c:legendEntry>` overrides off the chart. Returns
- * `undefined` when the chart declares no entries so the field is
- * elided entirely on a clean parse — absence and an empty array
- * round-trip identically through {@link cloneChart} (the writer skips
- * emission when the resolved list is empty).
- *
- * Each entry is admitted only when its `<c:idx val=".."/>` selector
- * parses to a non-negative integer (matches the OOXML
- * `xsd:unsignedInt` schema). Entries without an `<c:idx>` child or with
- * a malformed `val` attribute are dropped rather than surface a
- * fabricated index. The `<c:delete>` flag accepts the OOXML truthy /
- * falsy spellings (`"1"` / `"true"` / `"0"` / `"false"`); absence
- * collapses to `false` (the OOXML default — the entry renders).
- *
- * The caller is expected to confirm a visible legend exists before
- * invoking this — `<c:legendEntry>` only renders inside `<c:legend>`,
- * so reading from a chart that hides or omits the legend would surface
- * overrides with no on-screen effect.
- *
- * Duplicate `idx` values keep the first occurrence — Excel's renderer
- * treats later duplicates as overrides on the same series, but the
- * writer's `resolveLegendEntries` deduplicates with last-wins semantics
- * to give clone-through callers a way to override without manually
- * pruning. Reading "first wins" pairs naturally with that behaviour:
- * a parsed list re-emits cleanly, and an explicit clone override that
- * appends an entry still beats the parsed value.
- */
-function parseLegendEntries(chartEl: XmlElement): ChartLegendEntry[] | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-
-  const seen = new Set<number>();
-  const out: ChartLegendEntry[] = [];
-  for (const child of childElements(legend)) {
-    if (child.local !== "legendEntry") continue;
-    const idxEl = findChild(child, "idx");
-    if (!idxEl) continue;
-    const raw = idxEl.attrs.val;
-    if (typeof raw !== "string") continue;
-    const idx = Number.parseInt(raw, 10);
-    if (!Number.isFinite(idx) || idx < 0) continue;
-    if (seen.has(idx)) continue;
-    seen.add(idx);
-
-    const deleteEl = findChild(child, "delete");
-    const deleteFlag = deleteEl !== undefined ? readBoolVal(deleteEl.attrs.val) === true : false;
-    out.push({ idx, delete: deleteFlag });
-  }
-
-  return out.length > 0 ? out : undefined;
-}
-
-/**
- * Pull `<c:legend><c:txPr><a:p><a:pPr><a:defRPr sz="N"/></a:pPr></a:p>
- * </c:txPr></c:legend>` off the chart. Returns the font size in points
- * (range `1..400`).
- *
- * The OOXML `sz` attribute is in 100ths of a point — the reader
- * converts to points and rounds to the nearest 0.5pt (Excel's UI
- * exposes the same 0.5pt granularity). Absence of the element /
- * attribute and out-of-range / non-numeric / non-finite values all
- * collapse to `undefined` so a fresh chart and a chart that pinned an
- * out-of-range size both round-trip to the writer's "skip the size
- * attribute" path.
- *
- * Returns `undefined` whenever the chart omits the `<c:legend>` element
- * — there is no `<c:txPr>` slot to surface the size from in that case.
- * The `<a:defRPr>` lives inside `<c:txPr><a:p><a:pPr>` per the
- * CT_TextBody schema (the default-paragraph properties on the
- * legend's text-body's first paragraph); the lookup is scoped to that
- * path so a stray `<a:defRPr>` elsewhere in the chart (e.g. on the
- * chart title or an axis) cannot leak in. Mirrors the chart-title /
- * axis-title / axis tick-label readers exactly so a parsed value slots
- * straight back into the writer's emit path.
- */
-function parseLegendFontSize(chartEl: XmlElement): number | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  const txPr = findChild(legend, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.sz;
-  if (typeof raw !== "string") return undefined;
-  const trimmed = raw.trim();
-  if (trimmed.length === 0) return undefined;
-  const parsed = Number.parseInt(trimmed, 10);
-  if (!Number.isFinite(parsed)) return undefined;
-  // Convert from 100ths of a point to points, rounding to the nearest
-  // 0.5pt to match the granularity Excel's UI exposes. `Math.round`
-  // on `2 * (parsed / 100)` and dividing by 2 gives a clean half-step
-  // band that mirrors the writer's emit-time normalization. The
-  // chart-title / axis-title / tick-label sibling parsers use the
-  // identical conversion so a parsed value flows through every
-  // typography slot without bookkeeping the units.
-  const halfSteps = Math.round((parsed / TITLE_FONT_SZ_PER_POINT) * 2);
-  const points = halfSteps / 2;
-  if (points < TITLE_FONT_SIZE_MIN_PT || points > TITLE_FONT_SIZE_MAX_PT) return undefined;
-  return points;
-}
-
-/**
- * Pull `<c:legend><c:txPr><a:p><a:pPr><a:defRPr b=".."/></a:pPr></a:p>
- * </c:txPr></c:legend>` off the chart. Returns the bold flag.
- *
- * The OOXML `b` attribute is the `xsd:boolean` bold flag on
- * `CT_TextCharacterProperties` (ECMA-376 Part 1, §21.1.2.3.7). The
- * OOXML default `false` collapses to `undefined` so absence and
- * `b="0"` round-trip identically — only an explicit `b="1"` surfaces
- * `true`. Unknown / malformed `b` tokens drop to `undefined` rather
- * than fabricate a value the writer would never emit.
- *
- * Returns `undefined` whenever the chart omits the `<c:legend>`
- * element — there is no `<c:txPr>` slot to surface the flag from in
- * that case. The `<a:defRPr>` lives inside `<c:txPr><a:p><a:pPr>` per
- * the CT_TextBody schema (the default-paragraph properties on the
- * legend's text-body's first paragraph); the lookup is scoped to that
- * path so a stray `<a:defRPr>` elsewhere in the chart (e.g. on the
- * chart title or an axis) cannot leak in. Mirrors the chart-title /
- * axis-title / axis tick-label readers exactly so a parsed value
- * slots straight back into the writer's emit path.
- */
-function parseLegendBold(chartEl: XmlElement): boolean | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  const txPr = findChild(legend, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.b;
-  // OOXML `xsd:boolean` accepts `"1"` / `"true"` (truthy) and `"0"` /
-  // `"false"` (falsy). Truthy spellings surface `true`; falsy
-  // spellings collapse to `undefined` so the OOXML default and an
-  // explicit `b="0"` round-trip identically through `cloneChart`.
-  // Unknown / missing `b` tokens drop to `undefined` for the same
-  // reason — never fabricate a flag Excel would not emit.
-  if (raw === "1" || raw === "true") return true;
-  return undefined;
-}
-
-/**
- * Pull `<c:legend><c:txPr><a:p><a:pPr><a:defRPr i=".."/></a:pPr></a:p>
- * </c:txPr></c:legend>` off the chart. Returns the italic flag.
- *
- * The OOXML `i` attribute is the `xsd:boolean` italic flag on
- * `CT_TextCharacterProperties` (ECMA-376 Part 1, §21.1.2.3.7). The
- * OOXML default `false` collapses to `undefined` so absence and
- * `i="0"` round-trip identically — only an explicit `i="1"` surfaces
- * `true`. Unknown / malformed `i` tokens drop to `undefined` rather
- * than fabricate a value the writer would never emit.
- *
- * Returns `undefined` whenever the chart omits the `<c:legend>`
- * element — there is no `<c:txPr>` slot to surface the flag from in
- * that case. Mirrors the chart-title / axis-title / axis tick-label
- * readers exactly so a parsed value slots straight back into the
- * writer's emit path.
- */
-function parseLegendItalic(chartEl: XmlElement): boolean | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  const txPr = findChild(legend, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.i;
-  if (raw === "1" || raw === "true") return true;
-  return undefined;
-}
-
-/**
- * Pull `<c:legend><c:txPr><a:p><a:pPr><a:defRPr u=".."/></a:pPr></a:p>
- * </c:txPr></c:legend>` off the chart. Returns the underline flag.
- *
- * The OOXML `u` attribute is the `ST_TextUnderlineType` enumeration
- * on `CT_TextCharacterProperties`. Only `u="sng"` (Excel's UI variant
- * — single underline) surfaces `true`; the OOXML default `"none"`
- * (and every other variant the schema allows — `"dbl"`, `"heavy"`,
- * `"dotted"`, `"dotDash"`, `"wavy"`, etc.) collapse to `undefined` so
- * absence and `u="none"` round-trip identically through `cloneChart`.
- * Reporting any non-`"sng"` underline as `true` would silently
- * downgrade the choice to a single line on round-trip; the writer
- * emits only `u="sng"` / `u="none"`, matching the boolean shape the
- * UI exposes.
- *
- * Returns `undefined` whenever the chart omits the `<c:legend>`
- * element. Mirrors the chart-title / axis-title / axis tick-label
- * underline readers exactly so a parsed value slots straight back
- * into the writer's emit path.
- */
-function parseLegendUnderline(chartEl: XmlElement): boolean | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  const txPr = findChild(legend, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.u;
-  if (raw === "sng") return true;
-  return undefined;
-}
-
-/**
- * Pull `<c:legend><c:txPr><a:p><a:pPr><a:defRPr strike=".."/></a:pPr>
- * </a:p></c:txPr></c:legend>` off the chart. Returns the strikethrough
- * flag.
- *
- * The OOXML `strike` attribute is the `ST_TextStrikeType` enumeration
- * on `CT_TextCharacterProperties` — `"noStrike"` (default),
- * `"sngStrike"` (single line, Excel's UI variant), `"dblStrike"`
- * (double line, non-UI). Only `strike="sngStrike"` surfaces `true`;
- * the OOXML default `"noStrike"` and the double-line variant both
- * collapse to `undefined` so absence and `strike="noStrike"`
- * round-trip identically through `cloneChart`. Reporting `"dblStrike"`
- * as `true` would silently downgrade the choice to a single line on
- * re-emit; the writer emits only `"sngStrike"`, matching the boolean
- * shape the UI exposes.
- *
- * Returns `undefined` whenever the chart omits the `<c:legend>`
- * element. Mirrors the chart-title / axis-title / axis tick-label
- * strikethrough readers exactly so a parsed value slots straight back
- * into the writer's emit path.
- */
-function parseLegendStrikethrough(chartEl: XmlElement): boolean | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  const txPr = findChild(legend, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const raw = defRPr.attrs.strike;
-  if (raw === "sngStrike") return true;
-  return undefined;
-}
-
-/**
- * Pull the legend font color off the canonical
- * `<c:legend><c:txPr><a:p><a:pPr><a:defRPr><a:solidFill><a:srgbClr
- * val="RRGGBB"/></a:solidFill></a:defRPr></a:pPr></a:p></c:txPr>
- * </c:legend>` chain Excel writes when the user pins a custom font
- * color on the legend.
- *
- * Returns the 6-character uppercase hex string when the parser walks
- * the full chain and lands on an `<a:srgbClr val="RRGGBB"/>`. Theme
- * references (`<a:schemeClr>`), `<a:hslClr>`, `<a:sysClr>`, and
- * `<a:prstClr>` all collapse to `undefined` — only the literal RGB
- * triple round-trips losslessly through {@link writeChart}. Malformed
- * `val` tokens (wrong length, non-hex characters) likewise drop to
- * `undefined` rather than fabricate a value the writer would round-
- * trip into a malformed `<a:srgbClr>`.
- *
- * Returns `undefined` whenever the chart omits the `<c:legend>`
- * element or the canonical `<c:txPr><a:p><a:pPr><a:defRPr>
- * <a:solidFill><a:srgbClr>` chain is malformed at any link. Mirrors
- * the chart-title / axis-title / axis tick-label color readers
- * exactly so a parsed value slots straight back into the writer's
- * emit path.
- */
-function parseLegendFontColor(chartEl: XmlElement): string | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  const txPr = findChild(legend, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const solidFill = findChild(defRPr, "solidFill");
-  if (!solidFill) return undefined;
-  const srgbClr = findChild(solidFill, "srgbClr");
-  if (!srgbClr) return undefined;
-  return normalizeRgbHex(srgbClr.attrs.val);
-}
-
-/**
- * Pull the legend font family off the canonical
- * `<c:legend><c:txPr><a:p><a:pPr><a:defRPr><a:latin typeface=".."/>
- * </a:defRPr></a:pPr></a:p></c:txPr></c:legend>` chain Excel writes
- * when the user pins a typeface on the legend.
- *
- * The OOXML `<a:latin>` element carries the typeface name on
- * `CT_TextFont` (ECMA-376 Part 1, §21.1.2.3.7). The reader trims
- * surrounding whitespace and reports the trimmed typeface; empty /
- * whitespace-only `typeface` attributes and missing `<a:latin>`
- * elements both collapse to `undefined` so absence and the empty
- * form round-trip identically through the writer. Non-string
- * `typeface` tokens (defensive — the XML parser only ever surfaces
- * strings) likewise drop to `undefined`.
- *
- * Returns `undefined` whenever the chart omits the `<c:legend>`
- * element or the canonical `<c:txPr><a:p><a:pPr><a:defRPr><a:latin>`
- * chain is malformed at any link. Mirrors the chart-title /
- * axis-title / axis tick-label font family readers exactly so a
- * parsed value slots straight back into the writer's emit path.
- */
-function parseLegendFontFamily(chartEl: XmlElement): string | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  const txPr = findChild(legend, "txPr");
-  if (!txPr) return undefined;
-  const p = findChild(txPr, "p");
-  if (!p) return undefined;
-  const pPr = findChild(p, "pPr");
-  if (!pPr) return undefined;
-  const defRPr = findChild(pPr, "defRPr");
-  if (!defRPr) return undefined;
-  const latin = findChild(defRPr, "latin");
-  if (!latin) return undefined;
-  const raw = latin.attrs.typeface;
-  if (typeof raw !== "string") return undefined;
-  const trimmed = raw.trim();
-  if (trimmed.length === 0) return undefined;
-  return trimmed;
-}
-
-/**
- * Pull `<c:legend><c:layout><c:manualLayout>` off the chart. Reflects
- * Excel's "Format Legend -> Position -> Custom" knob — the `(x, y)`
- * anchor and `(w, h)` size of the legend block as fractions of the
- * chart frame in the `0..1` band.
- *
- * The OOXML schema (`CT_ManualLayout`, ECMA-376 Part 1, §21.2.2.115)
- * exposes optional `<c:x>` / `<c:y>` / `<c:w>` / `<c:h>` children whose
- * `val` attributes carry an `xsd:double`. The reader admits the
- * coordinate only when `val` parses to a finite number in the `0..1`
- * band; out-of-range / non-finite / non-numeric tokens drop to
- * `undefined` on the matching axis so absence and a malformed token
- * round-trip identically through {@link cloneChart}.
- *
- * Both `<c:xMode val="edge"/>` (absolute fraction of the chart frame)
- * and `<c:xMode val="factor"/>` (delta from auto-layout) are accepted
- * — the reader surfaces the same `ChartManualLayout` shape regardless,
- * since the writer always normalizes to `"edge"` on emit (Excel itself
- * emits the absolute form when the user drags an element to a custom
- * position).
- *
- * Returns `undefined` whenever the chart omits the `<c:legend>` /
- * `<c:layout>` / `<c:manualLayout>` chain at any link, or when every
- * coordinate dropped on normalization — the field is omitted entirely
- * on a clean parse so absence and an empty layout round-trip identically
- * through the writer.
- */
-function parseLegendLayout(chartEl: XmlElement): ChartManualLayout | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  return parseManualLayout(legend);
-}
-
-/**
- * Pull the legend background fill color off the canonical
- * `<c:legend><c:spPr><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></c:spPr></c:legend>` chain Excel writes when the
- * user pins "Format Legend -> Fill -> Solid fill -> Color".
- *
- * Returns the 6-character uppercase hex string when the parser walks
- * the full chain and lands on an `<a:srgbClr val="RRGGBB"/>`. Theme
- * references (`<a:schemeClr>`), `<a:hslClr>`, `<a:sysClr>`, and
- * `<a:prstClr>` all collapse to `undefined` — only the literal RGB
- * triple round-trips losslessly through {@link writeChart}. Non-solid
- * fills (`<a:noFill>`, `<a:gradFill>`, `<a:pattFill>`, `<a:blipFill>`)
- * likewise drop to `undefined` so a round-trip never fabricates a
- * fill the writer cannot reproduce on emit. Malformed `val` tokens
- * (wrong length, non-hex characters) drop to `undefined` rather than
- * fabricate a value the writer would round-trip into a malformed
- * `<a:srgbClr>`.
- *
- * The lookup is scoped to direct children of `<c:legend>` so a stray
- * `<c:spPr>` elsewhere (e.g. on a series or on the legend's
- * `<c:txPr>` block — the latter is purely text-character-properties,
- * but a malformed source might place a `<c:spPr>` there in error)
- * cannot leak in. Returns `undefined` whenever the chart omits the
- * `<c:legend>` element or the `<c:spPr><a:solidFill><a:srgbClr>`
- * chain is malformed at any link. Mirrors the chart-title /
- * axis-title / axis tick-label / legend font color readers exactly
- * so a parsed value slots straight back into the writer's emit path.
- *
- * Independent of {@link parseLegendFontColor}: the fill lives on
- * `<c:legend><c:spPr>`, the font color lives on
- * `<c:legend><c:txPr>` — the two readers walk disjoint paths so a
- * caller can pin both knobs without conflict.
- */
-function parseLegendFillColor(chartEl: XmlElement): string | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  return parseSpPrFill(legend);
-}
-
-/**
- * Pull `<c:legend><c:spPr><a:ln><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></a:ln></c:spPr></c:legend>` off the legend block.
- * Returns the line stroke color as a 6-character uppercase hex string.
- *
- * The OOXML `<a:srgbClr>` element carries the literal sRGB color
- * (`CT_SRgbColor`, ECMA-376 Part 1, §20.1.2.3.32) inside the line's
- * solid fill choice (`CT_LineProperties`, §20.1.2.3.24) which itself
- * sits inside `<c:spPr>` (`CT_ShapeProperties`, §20.1.2.3.13). The
- * `<c:spPr>` slot lives between `<c:overlay>` and `<c:txPr>` per
- * CT_Legend (ECMA-376 Part 1, §21.2.2.114).
- *
- * The reader surfaces only the literal `<a:srgbClr>` form — absence,
- * non-solid line fills (`<a:noFill>` / `<a:gradFill>` /
- * `<a:pattFill>`), and theme-color references (`<a:schemeClr>`) all
- * collapse to `undefined` so a chart that pinned a stroke the writer
- * cannot reproduce on emit drops the field rather than fabricate one
- * Excel would render differently. Malformed `val` tokens (wrong
- * length, non-hex characters, alpha-channel forms, non-string
- * escapes) likewise drop to `undefined`.
- *
- * The lookup is scoped to direct children of `<c:legend>` so a stray
- * `<c:spPr>` elsewhere (e.g. on a series, on an axis, on the legend's
- * `<c:txPr>` block) cannot leak in. Returns `undefined` whenever the
- * chart omits the `<c:legend>` element or the `<c:spPr><a:ln>
- * <a:solidFill><a:srgbClr>` chain is malformed at any link. Mirrors
- * the writer-side {@link SheetChart.legendBorderColor} so a parsed
- * value slots straight into {@link cloneChart} without conversion.
- *
- * Independent of {@link parseLegendFillColor}: the stroke lives on
- * `<c:legend><c:spPr><a:ln><a:solidFill>`, the fill lives on
- * `<c:legend><c:spPr><a:solidFill>` — the two readers walk disjoint
- * children of the same `<c:spPr>` block so a caller can pin both
- * knobs without conflict.
- */
-function parseLegendBorderColor(chartEl: XmlElement): string | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  return parseSpPrBorderColor(legend);
-}
-
-/**
- * Pull the `w` attribute off `<c:legend><c:spPr><a:ln w="EMU"/>` and
- * return the stroke width in points after clamping to the
- * `0.25..13.5` pt band Excel's UI exposes. The OOXML `w` attribute
- * carries the stroke width in English Metric Units (1 pt = 12 700 EMU)
- * per `CT_LineProperties` (ECMA-376 Part 1, §20.1.2.3.24); the reader
- * snaps the result to the 0.25 pt grid so a parsed-then-written width
- * does not drift across round-trips (Excel rounds in the UI anyway).
- *
- * Returns `undefined` when the chart omits `<c:legend>`, when the
- * legend has no `<c:spPr><a:ln w=..>` slot, when the attribute is
- * missing, when the value cannot be parsed as a finite positive
- * number, or when it parses to zero (Excel's "no border" marker — the
- * writer-side knob does not model that state). Mirrors the writer-side
- * {@link SheetChart.legendBorderWidth} so a parsed value slots
- * straight into {@link cloneChart} without conversion.
- *
- * The lookup is scoped to direct children of `<c:legend>` so a stray
- * `<a:ln w=..>` elsewhere (on a series stroke, on an axis line, on the
- * plot-area border) cannot leak in. Mirrors {@link parseLegendBorderColor} —
- * same `<c:spPr>` host on the same `<c:legend>` parent — but lands on
- * the `w` attribute rather than the `<a:solidFill><a:srgbClr>` color
- * child.
- */
-function parseLegendBorderWidth(chartEl: XmlElement): number | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  return parseBorderWidthFromSpPr(legend);
-}
-
-/**
- * Pull `<c:legend><c:spPr><a:ln><a:prstDash val=".."/></a:ln></c:spPr>
- * </c:legend>` off a chart. Returns the recognized
- * {@link ChartBorderDash} value when the chain is present and the
- * value is a valid `ST_PresetLineDashVal` token other than the OOXML
- * default `"solid"`. Returns `undefined` when the chart omits
- * `<c:legend>`, when the chain is broken, or when the value matches
- * the default. Mirrors {@link parseTitleBorderDash} on a different
- * host element.
- */
-function parseLegendBorderDash(chartEl: XmlElement): ChartBorderDash | undefined {
-  const legend = findChild(chartEl, "legend");
-  if (!legend) return undefined;
-  return parseBorderDashFromSpPr(legend);
-}
 
 
 /**
