+++ src/xlsx/chart-writer.ts
@@ -60,6 +60,29 @@ import {
   TXPR_ROT_PER_DEGREE,
 } from "./chart/text";
 import {
+  type LegendPos,
+  type ResolvedLegendEntry,
+  buildLegend,
+  buildLegendSpPr,
+  buildLegendTxPr,
+  normalizeLegendFontFamily,
+  resolveLegendBold,
+  resolveLegendBorderColor,
+  resolveLegendBorderDash,
+  resolveLegendBorderWidth,
+  resolveLegendEntries,
+  resolveLegendFillColor,
+  resolveLegendFontColor,
+  resolveLegendFontFamily,
+  resolveLegendFontSize,
+  resolveLegendItalic,
+  resolveLegendLayout,
+  resolveLegendOverlay,
+  resolveLegendPosition,
+  resolveLegendStrikethrough,
+  resolveLegendUnderline,
+} from "./chart/legend";
+import {
   buildTitle,
   buildTitleSpPr,
   normalizeTitleBold,
@@ -4985,658 +5008,7 @@ function buildDataLabelsTxPr(
   ]);
 }
 
-// ── Legend ───────────────────────────────────────────────────────────
-
-type LegendPos = "t" | "b" | "l" | "r" | "tr";
-
-function resolveLegendPosition(chart: SheetChart): LegendPos | null {
-  if (chart.legend === false) return null;
-  if (chart.legend === undefined) {
-    // Sensible defaults that match Excel's behaviour.
-    return chart.type === "scatter" ? "b" : "r";
-  }
-  switch (chart.legend) {
-    case "top":
-      return "t";
-    case "bottom":
-      return "b";
-    case "left":
-      return "l";
-    case "right":
-      return "r";
-    case "topRight":
-      return "tr";
-  }
-}
-
-function buildLegend(
-  pos: LegendPos,
-  overlay: boolean,
-  entries: readonly ResolvedLegendEntry[],
-  fontSizePt: number | undefined,
-  bold: boolean | undefined,
-  italic: boolean | undefined,
-  underline: boolean | undefined,
-  strikethrough: boolean | undefined,
-  rgbHex: string | undefined,
-  fontFamily: string | undefined,
-  layout: ResolvedManualLayout | undefined,
-  fillRgbHex: string | undefined,
-  borderRgbHex: string | undefined,
-  borderWidthPt: number | undefined,
-  borderDash: ChartBorderDash | undefined,
-): string {
-  const children: string[] = [xmlSelfClose("c:legendPos", { val: pos })];
-
-  // CT_Legend sequence places `<c:legendEntry>` after `<c:legendPos>`
-  // and before `<c:layout>` / `<c:overlay>` (ECMA-376 Part 1,
-  // §21.2.2.114). Each entry is emitted with both `<c:idx>` and
-  // `<c:delete>` so a re-parse sees the canonical shape — Excel itself
-  // emits `<c:delete val="1"/>` whenever the action is "Hide legend
-  // entry", and the writer mirrors that even for the OOXML default
-  // `false` value (an explicit `<c:delete val="0"/>` round-trips
-  // cleanly through `parseChart`).
-  for (const entry of entries) {
-    children.push(
-      xmlElement("c:legendEntry", undefined, [
-        xmlSelfClose("c:idx", { val: entry.idx }),
-        xmlSelfClose("c:delete", { val: entry.delete ? 1 : 0 }),
-      ]),
-    );
-  }
-
-  // CT_Legend sequence places `<c:layout>` between `<c:legendEntry>`
-  // and `<c:overlay>` per ECMA-376 Part 1, §21.2.2.114. The writer
-  // skips emission entirely when the caller pinned no coordinates so a
-  // fresh chart matches Excel's reference serialization byte-for-byte
-  // (Excel itself omits the block when the legend renders at the
-  // auto-layout position). Each axis is independently optional so the
-  // helper drops `<c:x>` / `<c:y>` / `<c:w>` / `<c:h>` slots whose
-  // value did not survive normalization.
-  const layoutXml = buildManualLayout(layout);
-  if (layoutXml !== undefined) {
-    children.push(layoutXml);
-  }
-
-  children.push(xmlSelfClose("c:overlay", { val: overlay ? 1 : 0 }));
-
-  // CT_Legend sequence places `<c:spPr>` between `<c:overlay>` and
-  // `<c:txPr>` (ECMA-376 Part 1, §21.2.2.114). The writer skips
-  // emission entirely when the caller did not pin a fill / border
-  // color so a fresh chart matches Excel's reference serialization
-  // byte-for-byte — Excel itself omits the block whenever the legend
-  // renders at the theme default fill / stroke (typically a transparent
-  // legend background with no `<c:spPr>` block). The writer authors
-  // `<a:solidFill>` (fill) and `<a:ln>` (stroke) here in
-  // `CT_ShapeProperties` schema order; other `CT_ShapeProperties`
-  // children (`<a:effectLst>` effects, gradient / pattern / picture
-  // fills, line dash / width / compound styles) are not modelled at
-  // this layer.
-  const legendSpPrXml = buildLegendSpPr(fillRgbHex, borderRgbHex, borderWidthPt, borderDash);
-  if (legendSpPrXml !== undefined) {
-    children.push(legendSpPrXml);
-  }
-
-  // CT_Legend sequence places `<c:txPr>` after `<c:spPr>` (and before
-  // the optional `<c:extLst>`). The writer skips emission entirely
-  // when no typography knob is pinned so a fresh chart matches Excel's
-  // reference serialization byte-for-byte (Excel itself omits the
-  // block whenever the legend renders at the theme-default style).
-  // The block currently carries the legend font size, bold, italic,
-  // underline, strikethrough, font color, and font family knobs. The
-  // `<a:bodyPr>` carries no rotation attribute — the legend is not
-  // rotatable in Excel's UI, mirroring how the axis tick-label
-  // `<c:txPr>` slot drops `rot` when only typography knobs are pinned.
-  const txPrXml = buildLegendTxPr(
-    fontSizePt,
-    bold,
-    italic,
-    underline,
-    strikethrough,
-    rgbHex,
-    fontFamily,
-  );
-  if (txPrXml !== undefined) {
-    children.push(txPrXml);
-  }
-
-  return xmlElement("c:legend", undefined, children);
-}
-
-/**
- * Build the `<c:spPr>` element on `<c:legend>` that carries the
- * legend's background fill, border (line stroke) color, and border
- * width. Returns `undefined` when no knob is pinned so the caller can
- * elide the entire block — Excel's reference serialization omits
- * `<c:spPr>` from `<c:legend>` whenever the legend renders at the
- * theme default fill / stroke (typically a transparent legend
- * background with no `<c:spPr>` block).
- *
- * The emitted block mirrors the minimal `<c:spPr>` shape Excel writes
- * when the user pins "Format Legend -> Fill -> Solid fill -> Color"
- * and / or "Format Legend -> Border -> Solid line -> Color" / "Width":
- * `<c:spPr><a:solidFill><a:srgbClr val="RRGGBB"/></a:solidFill>
- * <a:ln w="EMU"><a:solidFill><a:srgbClr val="RRGGBB"/></a:solidFill>
- * </a:ln></c:spPr>`. The `val` attribute holds the canonical
- * 6-character uppercase hex form (the writer normalizes the input
- * ahead of this call so a malformed source value never reaches emit).
- * The width attribute lands on `<a:ln>` (EMU; 1 pt = 12 700 EMU)
- * authored together with the border-color child so a stroke-only or
- * color-only legend still emits a single `<a:ln>` block. When at least
- * one knob lands on the wire, the children are emitted in
- * `CT_ShapeProperties` schema order: `<a:solidFill>` (fill) then
- * `<a:ln>` (line / stroke).
- *
- * Mirrors the chart-title / plot-area / axis-title `<c:spPr>` slots
- * so a single hex string threads cleanly through every fill / stroke
- * knob the writer authors.
- */
-function buildLegendSpPr(
-  fillRgbHex: string | undefined,
-  borderRgbHex: string | undefined,
-  borderWidthPt: number | undefined,
-  borderDash: ChartBorderDash | undefined,
-): string | undefined {
-  if (
-    fillRgbHex === undefined &&
-    borderRgbHex === undefined &&
-    borderWidthPt === undefined &&
-    borderDash === undefined
-  ) {
-    return undefined;
-  }
-  const children: string[] = [];
-  if (fillRgbHex !== undefined) {
-    children.push(
-      xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: fillRgbHex })]),
-    );
-  }
-  if (borderRgbHex !== undefined || borderWidthPt !== undefined || borderDash !== undefined) {
-    const lnAttrs: Record<string, string | number> = {};
-    if (borderWidthPt !== undefined) {
-      // OOXML stores stroke width in EMU (1 pt = 12 700 EMU). Round to
-      // the nearest integer because the schema types `w` as `xsd:int`.
-      lnAttrs.w = Math.round(borderWidthPt * EMU_PER_PT);
-    }
-    const lnChildren: string[] = [];
-    if (borderRgbHex !== undefined) {
-      lnChildren.push(
-        xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: borderRgbHex })]),
-      );
-    }
-    // `<a:prstDash>` follows `<a:solidFill>` per CT_LineProperties
-    // schema sequence (ECMA-376 Part 1, §20.1.2.3.24).
-    if (borderDash !== undefined) {
-      lnChildren.push(xmlSelfClose("a:prstDash", { val: borderDash }));
-    }
-    children.push(
-      lnChildren.length === 0
-        ? xmlSelfClose("a:ln", lnAttrs)
-        : xmlElement("a:ln", Object.keys(lnAttrs).length > 0 ? lnAttrs : undefined, lnChildren),
-    );
-  }
-  return xmlElement("c:spPr", undefined, children);
-}
-
-/**
- * Build the `<c:txPr>` block that carries the legend's typography pins
- * (currently the font size, bold, and italic flags). Returns
- * `undefined` when every input is unset so the caller can elide the
- * element entirely (Excel's reference serialization omits `<c:txPr>`
- * from `<c:legend>` when the legend renders at the theme-default
- * style).
- *
- * The emitted block mirrors the minimal `<c:txPr>` shape Excel writes
- * when the user pins a legend typography knob — `<a:bodyPr/>` (no
- * rotation because the legend is not rotatable), `<a:lstStyle/>` is
- * the empty list-style placeholder the schema requires, and the
- * `<a:p><a:pPr><a:defRPr sz="N" b=".." i=".."/></a:pPr><a:endParaRPr/>
- * </a:p>` paragraph stub Excel always emits hosts the typography
- * attributes on `<a:defRPr>`. Mirrors the chart-title / axis-title /
- * tick-label `<c:txPr>` slots exactly so a re-parse picks the values
- * off the canonical default-paragraph slot every other typography
- * reader expects.
- *
- * The bold / italic flags emit literal `b="1"` / `b="0"` / `i="1"` /
- * `i="0"` whenever the input is a boolean — `false` pins the OOXML
- * default explicitly, which is functionally identical to absence but
- * lets a clone target override an upstream `1` from a templated chart.
- */
-function buildLegendTxPr(
-  fontSizePt: number | undefined,
-  bold: boolean | undefined,
-  italic: boolean | undefined,
-  underline: boolean | undefined,
-  strikethrough: boolean | undefined,
-  rgbHex: string | undefined,
-  fontFamily: string | undefined,
-): string | undefined {
-  if (
-    fontSizePt === undefined &&
-    bold === undefined &&
-    italic === undefined &&
-    underline === undefined &&
-    strikethrough === undefined &&
-    rgbHex === undefined &&
-    fontFamily === undefined
-  )
-    return undefined;
-  const defRPrAttrs: Record<string, string | number> = {};
-  if (fontSizePt !== undefined) defRPrAttrs.sz = fontSizePt * TITLE_FONT_SZ_PER_POINT;
-  if (bold !== undefined) defRPrAttrs.b = bold ? 1 : 0;
-  if (italic !== undefined) defRPrAttrs.i = italic ? 1 : 0;
-  if (underline !== undefined) defRPrAttrs.u = underline ? "sng" : "none";
-  // Strikethrough rides as `strike="sngStrike"` on the same
-  // `<a:defRPr>` slot. Absence collapses to omitting the attribute
-  // entirely (the OOXML default `"noStrike"` is functionally identical
-  // to absence — the reader collapses both to `undefined`). The writer
-  // never emits `"noStrike"` or `"dblStrike"` so the surfaced shape
-  // stays consistent with Excel's UI checkbox.
-  if (strikethrough === true) defRPrAttrs.strike = "sngStrike";
-  // OOXML's `<a:defRPr><a:solidFill><a:srgbClr val="RRGGBB"/>
-  // </a:solidFill></a:defRPr>` carries the legend font color.
-  // Absence (`undefined`) collapses to omitting the entire
-  // `<a:solidFill>` block so the legend inherits the theme text
-  // color (Excel's reference behavior for a fresh legend that has
-  // not had a custom font color picked).
-  const solidFillChild = rgbHex
-    ? xmlElement("a:solidFill", undefined, [xmlSelfClose("a:srgbClr", { val: rgbHex })])
-    : undefined;
-  // OOXML's `<a:defRPr><a:latin typeface=".."/></a:defRPr>` carries
-  // the legend font family. The `<a:latin>` element follows
-  // `<a:solidFill>` per the CT_TextCharacterProperties child sequence
-  // (ECMA-376 Part 1, §21.1.2.3.7). Absence (`undefined`) collapses
-  // to omitting the entire `<a:latin>` element so the legend inherits
-  // the theme typeface (Excel's reference behavior for a fresh legend
-  // that has not had a custom font picked).
-  const latinChild = fontFamily ? xmlSelfClose("a:latin", { typeface: fontFamily }) : undefined;
-  // When a fill color or a typeface is set the `<a:defRPr>` slot
-  // expands from self-closing to wrapping the children; otherwise the
-  // writer keeps the existing self-closing form so a fresh legend
-  // with no custom color or font matches Excel's reference
-  // serialization byte-for-byte. Children are emitted in
-  // CT_TextCharacterProperties' canonical schema order: solidFill
-  // first, then latin.
-  const defRPrChildren: string[] = [];
-  if (solidFillChild) defRPrChildren.push(solidFillChild);
-  if (latinChild) defRPrChildren.push(latinChild);
-  const defRPr =
-    defRPrChildren.length > 0
-      ? xmlElement("a:defRPr", defRPrAttrs, defRPrChildren)
-      : xmlSelfClose("a:defRPr", defRPrAttrs);
-  return xmlElement("c:txPr", undefined, [
-    xmlSelfClose("a:bodyPr"),
-    xmlSelfClose("a:lstStyle"),
-    xmlElement("a:p", undefined, [
-      xmlElement("a:pPr", undefined, [defRPr]),
-      xmlSelfClose("a:endParaRPr", { lang: "en-US" }),
-    ]),
-  ]);
-}
-
-/**
- * Resolve `<c:legend><c:txPr><a:p><a:pPr><a:defRPr sz="N"/></a:pPr>
- * </a:p></c:txPr></c:legend>` from {@link SheetChart.legendFontSize}.
- *
- * Returns the size in points (`1..400`), or `undefined` when the chart
- * leaves the field unset / passed an out-of-range or non-numeric token.
- * The flag is only meaningful when the chart actually emits a legend —
- * the caller is expected to gate the call on the resolved legend
- * visibility (`resolveLegendPosition` returning a non-null value), so a
- * chart that hides the legend silently drops the value rather than
- * emit a `<c:txPr>` block with no on-screen effect.
- *
- * Mirrors `resolveTitleFontSize` / `resolveAxisTitleFontSize` /
- * `resolveAxisLabelFontSize` exactly so a single configuration call
- * threads cleanly through every typography slot Excel exposes.
- */
-function resolveLegendFontSize(chart: SheetChart): number | undefined {
-  return normalizeTitleFontSize(chart.legendFontSize);
-}
-
-/**
- * Resolve `<c:legend><c:txPr><a:p><a:pPr><a:defRPr b=".."/></a:pPr>
- * </a:p></c:txPr></c:legend>` from {@link SheetChart.legendBold}.
- *
- * Returns the bold flag, or `undefined` when the chart leaves the
- * field unset / passed a non-boolean token. The flag is only
- * meaningful when the chart actually emits a legend — the caller is
- * expected to gate the call on the resolved legend visibility
- * (`resolveLegendPosition` returning a non-null value), so a chart
- * that hides the legend silently drops the value rather than emit a
- * `<c:txPr>` block with no on-screen effect.
- *
- * Mirrors `resolveTitleBold` / `resolveAxisTitleBold` /
- * `resolveAxisLabelBold` exactly — only literal `true` / `false` pass
- * through; non-boolean tokens collapse to `undefined` so the writer
- * drops the slot rather than emit a value Excel would reject.
- */
-function resolveLegendBold(chart: SheetChart): boolean | undefined {
-  const value = chart.legendBold;
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:legend><c:txPr><a:p><a:pPr><a:defRPr i=".."/></a:pPr>
- * </a:p></c:txPr></c:legend>` from {@link SheetChart.legendItalic}.
- *
- * Returns the italic flag, or `undefined` when the chart leaves the
- * field unset / passed a non-boolean token. The flag is only
- * meaningful when the chart actually emits a legend — the caller is
- * expected to gate the call on the resolved legend visibility.
- *
- * Mirrors `resolveTitleItalic` / `resolveAxisTitleItalic` /
- * `resolveAxisLabelItalic` exactly — only literal `true` / `false`
- * pass through; non-boolean tokens collapse to `undefined`.
- */
-function resolveLegendItalic(chart: SheetChart): boolean | undefined {
-  const value = chart.legendItalic;
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:legend><c:txPr><a:p><a:pPr><a:defRPr u=".."/></a:pPr>
- * </a:p></c:txPr></c:legend>` from {@link SheetChart.legendUnderline}.
- *
- * Returns the underline flag, or `undefined` when the chart leaves
- * the field unset / passed a non-boolean token. The flag is only
- * meaningful when the chart actually emits a legend — the caller is
- * expected to gate the call on the resolved legend visibility.
- *
- * Mirrors `resolveTitleUnderline` / `resolveAxisTitleUnderline` /
- * `resolveAxisLabelUnderline` exactly — only literal `true` / `false`
- * pass through; non-boolean tokens collapse to `undefined`. The
- * writer translates `true` into `u="sng"` (Excel's UI variant —
- * single underline) and `false` into `u="none"` at emit time.
- */
-function resolveLegendUnderline(chart: SheetChart): boolean | undefined {
-  const value = chart.legendUnderline;
-  if (value === true) return true;
-  if (value === false) return false;
-  return undefined;
-}
-
-/**
- * Resolve `<c:legend><c:txPr><a:p><a:pPr><a:defRPr strike=".."/></a:pPr>
- * </a:p></c:txPr></c:legend>` from
- * {@link SheetChart.legendStrikethrough}.
- *
- * Returns `true` when the chart pins the strikethrough flag literally;
- * every other value (explicit `false`, absence, non-boolean tokens
- * leaking past the type guard) collapses to `undefined` so the writer
- * never emits a `strike` attribute below `"sngStrike"`. The OOXML
- * default `"noStrike"` is functionally identical to absence — the
- * writer keeps the surfaced shape consistent with what Excel's UI
- * authors (`"sngStrike"` only, never `"noStrike"` or `"dblStrike"`),
- * mirroring how `resolveTitleStrike` lands on the title's `<a:defRPr>`.
- *
- * Collapsing `false` to `undefined` (instead of the `boolean`-pass-through
- * shape that `legendBold` / `legendItalic` / `legendUnderline` use)
- * mirrors how the chart-title / axis-title / axis tick-label
- * strikethrough writers also drop the attribute on `false`: a
- * standalone `legendStrikethrough: false` does not gratuitously
- * trigger the `<c:txPr>` block emission, so the legend stays at the
- * theme-default style and a fresh chart with no custom strikethrough
- * matches Excel's reference serialization byte-for-byte. The flag is
- * only meaningful when the chart actually emits a legend — the caller
- * is expected to gate the call on the resolved legend visibility.
- */
-function resolveLegendStrikethrough(chart: SheetChart): boolean | undefined {
-  if (chart.legendStrikethrough === true) return true;
-  return undefined;
-}
-
-/**
- * Resolve `<c:legend><c:txPr><a:p><a:pPr><a:defRPr><a:solidFill>
- * <a:srgbClr val="RRGGBB"/></a:solidFill></a:defRPr></a:pPr></a:p>
- * </c:txPr></c:legend>` from {@link SheetChart.legendFontColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the chart leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches the chart-title /
- * axis-title / axis tick-label color resolvers exactly. The fill is
- * only meaningful when the chart actually emits a legend — the
- * caller is expected to gate the call on the resolved legend
- * visibility. A chart whose legend is suppressed has no `<c:txPr>`
- * slot to host the fill in either case.
- */
-function resolveLegendFontColor(chart: SheetChart): string | undefined {
-  return normalizeTitleColor(chart.legendFontColor);
-}
-
-/**
- * Normalize a {@link SheetChart.legendFontFamily} value for the
- * `<c:legend><c:txPr><a:p><a:pPr><a:defRPr><a:latin typeface=".."/>
- * </a:defRPr></a:pPr></a:p></c:txPr></c:legend>` writer slot. Returns
- * the trimmed typeface string when the input is a non-empty string,
- * or `undefined` for any malformed token — empty / whitespace-only
- * strings, or non-string escapes from an untyped caller (`null`,
- * numbers, booleans, etc.).
- *
- * Absence and malformed tokens both collapse to `undefined` so the
- * writer skips the entire `<a:latin>` element and the legend inherits
- * the theme typeface (Excel's reference behavior for a fresh legend
- * without a custom font picked).
- */
-function normalizeLegendFontFamily(value: string | undefined): string | undefined {
-  if (typeof value !== "string") return undefined;
-  const trimmed = value.trim();
-  if (trimmed.length === 0) return undefined;
-  return trimmed;
-}
-
-/**
- * Resolve `<c:legend><c:txPr><a:p><a:pPr><a:defRPr><a:latin
- * typeface=".."/></a:defRPr></a:pPr></a:p></c:txPr></c:legend>` from
- * {@link SheetChart.legendFontFamily}.
- *
- * Returns the trimmed typeface string the writer emits, or
- * `undefined` when the chart leaves the field unset / passed an empty
- * or non-string token. Delegates to {@link normalizeLegendFontFamily}
- * so the accept-and-trim grammar matches the chart-title /
- * axis-title / axis tick-label font family resolvers exactly. The
- * element is only meaningful when the chart actually emits a legend —
- * the caller is expected to gate the call on the resolved legend
- * visibility. A chart whose legend is suppressed has no `<c:txPr>`
- * slot to host the typeface in either case.
- */
-function resolveLegendFontFamily(chart: SheetChart): string | undefined {
-  return normalizeLegendFontFamily(chart.legendFontFamily);
-}
-
-interface ResolvedLegendEntry {
-  idx: number;
-  delete: boolean;
-}
-
-/**
- * Normalize {@link SheetChart.legendEntries} into an emit-ready list.
- *
- * The OOXML schema (`CT_LegendEntry`) places `<c:idx val="N"/>` as the
- * required selector and `<c:delete val=".."/>` as the hide flag. Hucre
- * accepts a free-form `ChartLegendEntry[]` from callers; this helper
- * strips entries whose `idx` cannot land on a real series and
- * deduplicates duplicate `idx` values so the writer never emits the
- * same selector twice (the last entry wins so a clone-through that
- * appends an override naturally beats the source's parsed value).
- *
- * Validation rules:
- *   - `idx` must be a non-negative integer (matches `xsd:unsignedInt`
- *     on `<c:idx val=".."/>`); non-finite, negative, or non-integer
- *     values drop entirely rather than emit a token Excel rejects.
- *   - `delete` collapses to a strict boolean — anything other than
- *     literal `true` resolves to `false`. Mirrors how `legendOverlay`
- *     / `roundedCorners` / `plotVisOnly` treat their inputs.
- *
- * Entries are emitted in ascending `idx` order so the rendered chart
- * matches Excel's reference serialization (Excel sorts by `<c:idx>` on
- * write even when the in-memory list arrived unsorted). Returns an
- * empty array when the chart has no entries to emit so the caller can
- * avoid touching the legend block.
- */
-function resolveLegendEntries(chart: SheetChart): ResolvedLegendEntry[] {
-  const raw = chart.legendEntries;
-  if (!Array.isArray(raw) || raw.length === 0) return [];
-
-  const byIdx = new Map<number, ResolvedLegendEntry>();
-  for (const entry of raw) {
-    if (!entry || typeof entry !== "object") continue;
-    const idx = entry.idx;
-    if (typeof idx !== "number" || !Number.isFinite(idx)) continue;
-    if (!Number.isInteger(idx) || idx < 0) continue;
-    byIdx.set(idx, { idx, delete: entry.delete === true });
-  }
-
-  return Array.from(byIdx.values()).sort((a, b) => a.idx - b.idx);
-}
-
-/**
- * Resolve `<c:legend><c:overlay val=".."/></c:legend>` from
- * {@link SheetChart.legendOverlay}.
- *
- * Defaults to `false` (the OOXML default Excel itself emits — the
- * legend reserves its own slot and the plot area shrinks to make room).
- * Anything other than literal `true` collapses to `false` so a stray
- * non-boolean leaking through the type guard (e.g. `0` / `1` / `"true"`
- * / `null`) never produces `<c:overlay val="1"/>`. This matches how
- * `roundedCorners` / `plotVisOnly` / axis `hidden` treat their inputs:
- * a literal boolean is the only path to a non-default value.
- *
- * The writer always emits `<c:overlay>` because Excel's reference
- * serialization includes the element on every visible legend; only the
- * `val` flips when the caller pins `legendOverlay: true`.
- */
-function resolveLegendOverlay(chart: SheetChart): boolean {
-  return chart.legendOverlay === true;
-}
-
 // ── Manual Layout ────────────────────────────────────────────────────
-
-/**
- * Resolve `<c:legend><c:layout><c:manualLayout>...</c:manualLayout>
- * </c:layout></c:legend>` from {@link SheetChart.legendLayout}.
- *
- * Returns the normalized coordinate set, or `undefined` when every
- * axis the caller pinned dropped to `undefined` (so the writer can
- * elide the entire `<c:layout>` block — Excel's reference
- * serialization omits the element when the legend renders at the
- * auto-layout position). The element is only meaningful when the chart
- * actually emits a legend — the caller is expected to gate the call on
- * the resolved legend visibility.
- *
- * Coordinates outside the OOXML `0..1` band, `NaN`, `Infinity`, and
- * non-numeric inputs all collapse to `undefined` on the matching axis
- * so the writer drops the matching `<c:x>` / `<c:y>` / `<c:w>` /
- * `<c:h>` slot rather than emit a token Excel would reject.
- */
-function resolveLegendLayout(chart: SheetChart): ResolvedManualLayout | undefined {
-  return normalizeManualLayout(chart.legendLayout);
-}
-
-/**
- * Resolve `<c:legend><c:spPr><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></c:spPr></c:legend>` from
- * {@link SheetChart.legendFillColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the chart leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches the chart-title /
- * axis-title / axis tick-label / legend font color resolvers exactly.
- * The fill is only meaningful when the chart actually emits a
- * legend — the caller is expected to gate the call on the resolved
- * legend visibility (`resolveLegendPosition` returning a non-null
- * value). A chart whose legend is suppressed has no `<c:legend>`
- * block to host the `<c:spPr>` slot in either case.
- *
- * Independent of {@link resolveLegendFontColor}: the fill lands on
- * `<c:legend><c:spPr>`, the font color lands on
- * `<c:legend><c:txPr>` — the two resolvers target different children
- * of `<c:legend>` so a single configuration call can pin both.
- */
-function resolveLegendFillColor(chart: SheetChart): string | undefined {
-  return normalizeTitleColor(chart.legendFillColor);
-}
-
-/**
- * Resolve `<c:legend><c:spPr><a:ln><a:solidFill><a:srgbClr val="RRGGBB"/>
- * </a:solidFill></a:ln></c:spPr></c:legend>` from
- * {@link SheetChart.legendBorderColor}.
- *
- * Returns the 6-character uppercase hex string the writer emits, or
- * `undefined` when the chart leaves the field unset / passed a
- * malformed token. Delegates to {@link normalizeTitleColor} so the
- * accept-with-or-without-`#` grammar matches every other `<a:srgbClr>`
- * fill / line slot exactly. The stroke is only meaningful when the
- * chart actually emits a legend — the caller is expected to gate the
- * call on the resolved legend visibility (`resolveLegendPosition`
- * returning a non-null value). A chart whose legend is suppressed has
- * no `<c:legend>` block to host the `<c:spPr>` slot in either case.
- *
- * Independent of {@link resolveLegendFillColor}: the fill lands on
- * `<c:legend><c:spPr><a:solidFill>`, the stroke lands on
- * `<c:legend><c:spPr><a:ln><a:solidFill>` — the two resolvers feed
- * disjoint children of `<c:spPr>` so a single configuration call can
- * pin both. Mirrors the chart-title / axis-title / chart-space /
- * plot-area `<c:spPr>` slots — same hex grammar, same `<a:ln>` slot
- * on the `CT_ShapeProperties` schema — but lands on `<c:legend>`'s
- * own `<c:spPr>` block.
- */
-function resolveLegendBorderColor(chart: SheetChart): string | undefined {
-  return normalizeTitleColor(chart.legendBorderColor);
-}
-
-/**
- * Resolve `<c:legend><c:spPr><a:ln w="EMU"/></c:spPr></c:legend>` from
- * {@link SheetChart.legendBorderWidth}.
- *
- * Returns the point value clamped to the `0.25..13.5` pt band Excel's
- * UI exposes and snapped to the 0.25 pt grid, or `undefined` when the
- * chart leaves the field unset / passed a malformed token (`NaN`,
- * `Infinity`, non-finite). Delegates to {@link clampStrokeWidthPt} so
- * the snap / clamp grammar matches every other `<a:ln w=..>` slot the
- * writer authors (the series stroke knob `series[i].stroke.width`,
- * the plot-area border width knob {@link SheetChart.plotAreaBorderWidth}).
- * The width is only meaningful when the chart actually emits a
- * legend — the caller is expected to gate the call on the resolved
- * legend visibility (`resolveLegendPosition` returning a non-null
- * value). A chart whose legend is suppressed has no `<c:legend>`
- * block to host the `<c:spPr>` slot in either case.
- *
- * Independent of {@link resolveLegendBorderColor}: both knobs land on
- * the same `<a:ln>` element but on a different slot (the color child
- * `<a:solidFill>` versus the line's `w` attribute). Mirrors the
- * chart-title / axis-title / chart-space / plot-area `<c:spPr>` slots —
- * same EMU encoding, same `<a:ln>` host — but lands on `<c:legend>`'s
- * own `<c:spPr>` block.
- */
-function resolveLegendBorderWidth(chart: SheetChart): number | undefined {
-  return clampStrokeWidthPt(chart.legendBorderWidth);
-}
-
-/**
- * Resolve `<c:legend><c:spPr><a:ln><a:prstDash val=".."/></a:ln></c:spPr>
- * </c:legend>` from {@link SheetChart.legendBorderDash}.
- *
- * Returns the recognized {@link ChartBorderDash} value, or `undefined`
- * for the OOXML default `"solid"` and every unrecognized token —
- * delegates to {@link normalizeBorderDash} so absence and the OOXML
- * default round-trip identically. The dash is only meaningful when the
- * chart actually emits a legend; the caller gates this on the resolved
- * legend visibility.
- *
- * Independent of {@link resolveLegendBorderColor} and
- * {@link resolveLegendBorderWidth}: all three knobs land on the same
- * `<a:ln>` element but on different children / attributes.
- */
-function resolveLegendBorderDash(chart: SheetChart): ChartBorderDash | undefined {
-  return normalizeBorderDash(chart.legendBorderDash);
-}
 
 
 /**
